refactor: Split pontos-release file: Detach argparser and main command from big release.py file
