+++ pontos/release/__init__.py
@@ -16,6 +16,6 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from .release import main
+from .main import main
 
 __all__ = ['main']
